# Conflicts:
#   backend/eco2/src/main/java/com/web/eco2/model/repository/post/PostRepository.java

+++ frontend/eco2/src/components/dailyMission/missionItem/missionMainItem.jsx
@@ -1,5 +1,5 @@
 // React
-import React, { useState } from "react";
+import React, { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
 import { useDispatch } from "react-redux/es/exports";
 
@@ -9,14 +9,9 @@ import { deleteMission, clearMission } from "../../../store/mission/missionMainS
 // CSS
 import styles from "./dailyMission.module.css";
 
-const MissionMain = ({ idTest, content, trashCnt, missionIdTest, missionFlag }) => {
-  const [trash, setTrash] = useState(true); // 미션 인증 완료 시 쓰레기통 변화 하는 State
-  const [doitMission, setDoitMission] = useState(true); // 미션 인증 하기 위한 State
-  const [clearMissionT, setClearMission] = useState(false); // 미션 인증 완료 시 동그라미 색 변화 State
-
-  const clearMissionType = !clearMissionT ? "fa-regular fa-circle" : "fa-solid fa-circle"; // 미션 인증 색 변화
+const MissionMain = ({ idTest, content, missionIdTest, missionFlag }) => {
   const flagType = missionFlag ? "fa-solid fa-circle" : "fa-regular fa-circle"; // 미션 인증 색 변화
-  const trashType = trash ? styles.greenTrash : styles.whiteTrash; // 쓰레기통 색 변화
+  const realTrashType = missionFlag ? styles.whiteTrash : styles.greenTrash; // 미션 여부에 따른 쓰레기통
 
   const naviGate = useNavigate();
   const dispatch = useDispatch();
@@ -25,12 +20,17 @@ const MissionMain = ({ idTest, content, trashCnt, missionIdTest, missionFlag }) 
    * user 추가 미션 삭제 클릭 시 삭제를 위한 함수
    */
   const onDeleButton = () => {
-    if (window.confirm("미션을 삭제 하시겠습니까?")) {
-      dispatch(deleteMission({ id: idTest, missionId: missionIdTest }));
-      alert("삭제 완료!");
-      window.location.replace("/dailymissionMain");
+    if (!missionFlag) {
+      if (window.confirm("미션을 삭제 하시겠습니까?")) {
+        dispatch(deleteMission({ id: idTest, missionId: missionIdTest }));
+        alert("삭제 완료!");
+        window.location.replace("/dailymissionMain");
+      } else {
+        alert("포기하지 말고 화이팅!");
+      }
     } else {
-      alert("포기하지 말고 화이팅!");
+      // dispatch(deleteMission({ id: idTest, missionId: missionIdTest }));
+      alert("미션 완료시 삭제 불가합니다!");
     }
   };
 
@@ -38,39 +38,29 @@ const MissionMain = ({ idTest, content, trashCnt, missionIdTest, missionFlag }) 
    * 미션 완료 동그라미 클릭 시 인증게시글 이동하기 위한 함수.
    */
   const boolCheck = () => {
-    console.log(doitMission);
-    if (doitMission) {
+    if (!missionFlag) {
       if (window.confirm("미션완료!! 인증게시글 업로드 하시겠어요?")) {
-        naviGate("/post");
+        dispatch(clearMission({ id: idTest, missionId: missionIdTest }));
+        naviGate("/post", { customId: idTest });
       } else {
-        alert("인증게시글 업로드도 부탁해요 ㅠㅠ!");
+        alert("인증게시글 업로드시 완료 인증됩니다!!");
       }
+    } else {
+      alert("완료한 미션입니다!");
     }
-  };
-
-  /**
-   * 동그라미 클릭 시 미션 완료 플래그 변환을 위한 함수
-   */
-  const onClear = (idTest, missionIdTest) => {
-    dispatch(clearMission({ id: idTest, missionId: missionIdTest }));
   };
 
   return (
     <div>
       <div className={styles.mainList}>
         <i
-          className={`${flagType} ${clearMissionType}`}
+          className={`${flagType}`}
           onClick={() => {
-            onClear(idTest, missionIdTest);
-            setClearMission(!clearMissionT);
-            setDoitMission(!doitMission);
-            setTrash(!trash);
-            trashCnt(!trash);
-            boolCheck();
+            boolCheck(idTest, missionIdTest);
           }}
         ></i>
         <span>{content}</span>
-        <i className={`${"fa-solid fa-trash-can"} ${trashType}`} onClick={onDeleButton}></i>
+        <i className={`${"fa-solid fa-trash-can"} ${realTrashType}`} onClick={onDeleButton}></i>
       </div>
     </div>
   );
